Created StockSpan.java
The span of a stock's price on a given day is defined as the maximum number of consecutive days (including the current day) for which the stock price was less than or equal to the price on the current day. In simple terms, it measures how many days a stock had a lower or equal price leading up to the current day.