Created CelebrityFinder.java
The Celebrity Problem involves finding a celebrity in a group of people. A celebrity is someone who is known by everyone in the group but doesn't know anyone. It can be represented using a square matrix where M[i][j] represents whether person i knows person j (1 for "knows" and 0 for "doesn't know"). The goal is to identify the celebrity efficiently.